cloud-functions/ingest/requirements.txt
google-cloud-pubsub
google-cloud-datastore
google-cloud-core
sodapy
firebase_admin